Backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
bcrypt==5.0.0
celery==5.4.0
Django==6.0.3
django-cors-headers==4.9.0
django-environ==0.13.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.1.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-decouple==3.8
redis==5.2.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
